feature: move `ggshield ignore` command to `ggshield secret ignore`
